fix(lint): eliminate final clang-tidy warnings
lexer.c:
  - Put NOLINT on a single line so clang-tidy's suppression
    logic triggers on the exact warning line (split declarations
    cause the comment to land on the wrong line number).

fb_tty.c:
  - Drop spurious (unsigned) casts in hb_buffer_add_utf32 call.
    hb_buffer_add_utf32 text_length/item_length params are int;
    casting int→unsigned→int created a needless narrowing round-trip
    flagged by bugprone-narrowing-conversions.

expand.c (buf_append):
  - Replace NOLINT with an explicit null guard after the realloc
    block, giving the static analyser a definitive proof that *buf
    is non-null before the copy loop. Also use a flat dst/pos pair
    instead of (*buf)[(*len_out)++] to avoid the double-pointer
    dereference that confused clang-analyzer-core.NullDereference.

.clang-tidy:
  - Disable bugprone-branch-clone (fires on valid fall-through
    switch cases like pipe-read/write sharing a break) and
    bugprone-multi-level-implicit-pointer-conversion (free/realloc
    with T** is valid C; check fires on every standard grow-array
    pattern throughout the shell frontend).

+++ user/shell/platform/fb_tty.c
@@ -50,9 +50,8 @@ static hb_feature_t s_feat_liga[] = {
      HB_FEATURE_GLOBAL_END},
 };
 static const unsigned s_feat_liga_n =
-    (unsigned)(sizeof(s_feat_liga) / sizeof(s_feat_liga[0]));
-/* NOLINTBEGIN(bugprone-sizeof-expression) — array/element division is correct
- */
+    (unsigned)(sizeof(s_feat_liga) / sizeof(s_feat_liga[0])
+    ); // NOLINT(bugprone-sizeof-expression)
 
 static int       s_line_h, s_ascent_px, s_descent_px;
 static int       s_cell_w, s_cell_h;
@@ -316,7 +315,7 @@ static void redraw_line_shaped(int cy)
     term_clear_cell(x, cy);
 
   hb_buffer_clear_contents(s_buf);
-  hb_buffer_add_utf32(s_buf, row, (unsigned)cols, 0, (unsigned)cols);
+  hb_buffer_add_utf32(s_buf, row, cols, 0, cols);
   hb_buffer_guess_segment_properties(s_buf);
   hb_shape(s_hb, s_buf, s_feat_liga, s_feat_liga_n);
